Move each component in lib into separate folder
