--- desktop/packages/mullvad-vpn/src/renderer/lib/components/atoms/index.ts
@@ -1,1 +1,0 @@
-export * from './logo/Logo';
