+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/button/index.ts
@@ -1,0 +1,1 @@
+export * from './Button';
