+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/flex/index.ts
@@ -1,0 +1,1 @@
+export * from './Flex';
