+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/icon-button/index.ts
@@ -1,2 +1,1 @@
-export * from './Button';
 export * from './IconButton';
