+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/index.ts
@@ -1,5 +1,11 @@
+export * from './box';
+export * from './button';
+export * from './container';
+export * from './flex';
+export * from './icon-button';
 export * from './layout';
-export * from './atoms';
+export * from './logo';
+export * from './main-header';
+export * from './navigation-header';
+export * from './theme';
 export * from './typography';
-export * from './molecules';
-export * from './organisms';
