+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/layout/index.ts
@@ -1,5 +1,1 @@
-export * from './Box';
-export * from './Container';
-export * from './Flex';
 export * from './Layout';
-export * from './Theme';
