+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/main-header/components/MainHeaderIconButton.tsx
@@ -1,4 +1,4 @@
-import { IconButton, IconButtonProps } from '../../../molecules';
+import { IconButton, IconButtonProps } from '../../icon-button';
 
 export const MainHeaderIconButton = (props: IconButtonProps) => {
   return <IconButton variant="secondary" {...props} />;
