+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/navigation-header/NavigationHeader.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
-import { Colors, Spacings } from '../../../foundations';
-import { TransientProps } from '../../../types';
-import { Flex } from '../../layout';
+import { Colors, Spacings } from '../../foundations';
+import { TransientProps } from '../../types';
+import { Flex } from '../flex';
 import {
   NavigationHeaderButtonGroup,
   NavigationHeaderIconButton,
