+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/navigation-header/components/NavigationHeaderButtonGroup.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { Spacings } from '../../../../foundations';
-import { Flex } from '../../../layout';
+import { Spacings } from '../../../foundations';
+import { Flex } from '../../flex';
 
 export const NavigationHeaderButtonGroup = styled(Flex).attrs({
   $gap: Spacings.spacing6,
