+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/navigation-header/components/NavigationHeaderIconButton.tsx
@@ -1,4 +1,4 @@
-import { IconButton, IconButtonProps } from '../../../molecules';
+import { IconButton, IconButtonProps } from '../../icon-button';
 
 export const NavigationHeaderIconButton = (props: IconButtonProps) => {
   return <IconButton variant="secondary" {...props} />;
