+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/navigation-header/components/NavigationHeaderTitle.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { TitleMedium } from '../../../typography';
+import { TitleMedium } from '../../typography';
 import { useNavigationHeader } from '../NavigationHeaderContext';
 
 export interface NavigationHeaderTitleProps {
