--- desktop/packages/mullvad-vpn/src/renderer/lib/components/organisms/index.ts
@@ -1,2 +1,0 @@
-export * from './main-header';
-export * from './navigation-header';
