+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/theme/index.ts
@@ -1,0 +1,1 @@
+export * from './Theme';
